refactor: sql 파서 패키지 이동
유사한 역할을 하는 sql 파서끼리 동일한 패키지로 묶음.
