+++ src/main/java/codesquad/database/parser/InsertParser.java
@@ -1,4 +1,4 @@
-package codesquad.database;
+package codesquad.database.parser;
 
 import java.util.Arrays;
 import java.util.List;
